# metal: concurrently dispatch commands