Discussion see #2309

Function `ggml_metal_graph_find_concurrency` will write commands that can be issued concurrently to metal context `concur_list` array, when `ggml_metal_graph_compute` is called for the first time.

Tested on M1 Max 32c gpu with :
`./main -m model_file -n 256 -c 512 -s 123 -p "I believe the meaning of life is" --ignore-eos -ngl 1 --no-mmap -t 8`
|model| master | PR|improvement
|---|---|---|---
|33B Q4_0| 73.0 ms/tok | 70.1 ms/tok | ~4%
|7B Q4_0|  19.50 ms/tok | 18.26 ms/tok | ~6%

Btw, since we already have much faster kernels for both `Q_K` and `non Q_K`, is it a good time to do prompt evaluation on metal? Although current kernels are optimized for mat-vec multiplications, they still provide much better performance than cpu. (EDIT: not true for M1 Pro)